Extract descendant count out into its own component
