+++ public/app/features/browse-dashboards/components/BrowseActions/DeleteModal.tsx
@@ -2,12 +2,11 @@ import { css } from '@emotion/css';
 import React from 'react';
 
 import { GrafanaTheme2 } from '@grafana/data';
-import { Alert, ConfirmModal, Spinner, useStyles2 } from '@grafana/ui';
+import { ConfirmModal, useStyles2 } from '@grafana/ui';
 
-import { useGetAffectedItemsQuery } from '../../api/browseDashboardsAPI';
 import { DashboardTreeSelection } from '../../types';
 
-import { buildBreakdownString } from './utils';
+import { DescendantCount } from './DescendantCount';
 
 export interface Props {
   isOpen: boolean;
@@ -18,7 +17,6 @@ export interface Props {
 
 export const DeleteModal = ({ onConfirm, onDismiss, selectedItems, ...props }: Props) => {
   const styles = useStyles2(getStyles);
-  const { data, isFetching, isLoading, error } = useGetAffectedItemsQuery(selectedItems);
 
   const onDelete = () => {
     onConfirm();
@@ -29,14 +27,7 @@ export const DeleteModal = ({ onConfirm, onDismiss, selectedItems, ...props }: P
     <ConfirmModal
       body={
         <div className={styles.modalBody}>
-          This action will delete the following content:
-          <div className={styles.breakdown}>
-            <>
-              {data && buildBreakdownString(data.folder, data.dashboard, data.libraryPanel, data.alertRule)}
-              {(isFetching || isLoading) && <Spinner size={12} />}
-              {error && <Alert severity="error" title="Unable to retrieve descendant information" />}
-            </>
-          </div>
+          <DescendantCount selectedItems={selectedItems} label="This action will delete the following content:" />
         </div>
       }
       confirmationText="Delete"
@@ -50,11 +41,6 @@ export const DeleteModal = ({ onConfirm, onDismiss, selectedItems, ...props }: P
 };
 
 const getStyles = (theme: GrafanaTheme2) => ({
-  breakdown: css({
-    ...theme.typography.bodySmall,
-    color: theme.colors.text.secondary,
-    marginBottom: theme.spacing(2),
-  }),
   modalBody: css({
     ...theme.typography.body,
   }),
