+++ public/app/features/browse-dashboards/components/BrowseActions/MoveModal.tsx
@@ -1,14 +1,11 @@
-import { css } from '@emotion/css';
 import React, { useState } from 'react';
 
-import { GrafanaTheme2 } from '@grafana/data';
-import { Alert, Button, Field, Modal, Spinner, useStyles2 } from '@grafana/ui';
+import { Alert, Button, Field, Modal } from '@grafana/ui';
 import { FolderPicker } from 'app/core/components/Select/FolderPicker';
 
-import { useGetAffectedItemsQuery } from '../../api/browseDashboardsAPI';
 import { DashboardTreeSelection } from '../../types';
 
-import { buildBreakdownString } from './utils';
+import { DescendantCount } from './DescendantCount';
 
 export interface Props {
   isOpen: boolean;
@@ -19,9 +16,7 @@ export interface Props {
 
 export const MoveModal = ({ onConfirm, onDismiss, selectedItems, ...props }: Props) => {
   const [moveTarget, setMoveTarget] = useState<string>();
-  const styles = useStyles2(getStyles);
   const selectedFolders = Object.keys(selectedItems.folder).filter((uid) => selectedItems.folder[uid]);
-  const { data, isFetching, isLoading, error } = useGetAffectedItemsQuery(selectedItems);
 
   const onMove = () => {
     if (moveTarget !== undefined) {
@@ -33,14 +28,7 @@ export const MoveModal = ({ onConfirm, onDismiss, selectedItems, ...props }: Pro
   return (
     <Modal title="Move" onDismiss={onDismiss} {...props}>
       {selectedFolders.length > 0 && <Alert severity="warning" title="Moving this item may change its permissions." />}
-      This action will move the following content:
-      <div className={styles.breakdown}>
-        <>
-          {data && buildBreakdownString(data.folder, data.dashboard, data.libraryPanel, data.alertRule)}
-          {(isFetching || isLoading) && <Spinner size={12} />}
-          {error && <Alert severity="error" title="Unable to retrieve descendant information" />}
-        </>
-      </div>
+      <DescendantCount selectedItems={selectedItems} label="This action will move the following content:" />
       <Field label="Folder name">
         <FolderPicker allowEmpty onChange={({ uid }) => setMoveTarget(uid)} />
       </Field>
@@ -55,11 +43,3 @@ export const MoveModal = ({ onConfirm, onDismiss, selectedItems, ...props }: Pro
     </Modal>
   );
 };
-
-const getStyles = (theme: GrafanaTheme2) => ({
-  breakdown: css({
-    ...theme.typography.bodySmall,
-    color: theme.colors.text.secondary,
-    marginBottom: theme.spacing(2),
-  }),
-});
